experiments/calibration-preflight/pyproject.toml
[project]
name = "calibration-preflight"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.72.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "openai>=2.31.0",
    "scikit-learn>=1.8.0",
]
